posts are proper

+++ backend/social_distribution/views.py
@@ -3,7 +3,7 @@ import statistics
 from django.shortcuts import render
 from django.template import Context
 from rest_framework import viewsets, status
-from .serializers import PostSerializer, LoginSerializer, CommentSerializer, AuthorSerializer, InboxSerializer, InboxItemSerializer, RequestSerializer, FriendRequestSerializer, FollowersSerializer
+from .serializers import PostSerializer, LoginSerializer, CommentSerializer, AuthorSerializer, InboxSerializer, InboxItemSerializer, RequestSerializer, FriendRequestSerializer, FollowersSerializer, LikeSerializer
 from .models import Post, Author, Comment, Request, Inbox, InboxItem, FriendRequest, Followers
 from rest_framework.response import Response
 from rest_framework.views import APIView
@@ -15,13 +15,13 @@ from django.forms.models import model_to_dict
 from django.conf import settings
 import jwt
 from django.db.models import Q
+from rest_framework.decorators import api_view
 
-
-# Create your views here.
+# need to be changed to proper format
 class AuthorViewSet(APIView):
     def get(self, request, *args, **kwargs):
-        if (kwargs.get('pk')):
-            serializer = AuthorSerializer(Author.objects.filter(pk=kwargs.get('pk')),many=True) 
+        if (kwargs.get('id')):
+            serializer = AuthorSerializer(Author.objects.filter(id=kwargs.get('id')),many=True) 
         else:
             serializer = AuthorSerializer(Author.objects.all(),many=True) 
         return Response(serializer.data, status=status.HTTP_200_OK)
@@ -34,14 +34,14 @@ class InboxViewSet(APIView):
     def get(self, request):
         token = request.headers.get('Authorization', None)
         payload = LoginSerializer.validateToken(token)
-        pk = payload.get('user_id', None)
-        return inbox(request, pk)
+        id = payload.get('user_id', None)
+        return inbox(request, id)
     
     def delete(self, request):
         token = request.headers.get('Authorization', None)
         payload = LoginSerializer.validateToken(token)
-        pk = payload.get('user_id', None)
-        return inbox(request, pk)
+        id = payload.get('user_id', None)
+        return inbox(request, id)
 
 class HomeViewSet(APIView):
     def get(self, request):
@@ -53,7 +53,7 @@ class PostViewSet(APIView):
     def get(self, request):
         token = request.headers.get('Authorization', None)
         payload = LoginSerializer.validateToken(token)
-        author = Author.objects.get(pk = payload.get('user_id', None))
+        author = Author.objects.get(id = payload.get('user_id', None))
         posts = Post.objects.filter(author= author)
         serializer = PostSerializer(posts, many=True)
         return Response(serializer.data, status=status.HTTP_200_OK)
@@ -61,7 +61,7 @@ class PostViewSet(APIView):
     def post(self, request):
         token = request.headers.get('Authorization', None)
         payload = LoginSerializer.validateToken(token)
-        author = Author.objects.get(pk = payload.get('user_id', None))
+        author = Author.objects.get(id = payload.get('user_id', None))
         data = request.data
         data['author'] = author.id
         data['authorName'] = author.displayName
@@ -71,11 +71,11 @@ class PostViewSet(APIView):
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
     
-    def put(self, request, pk):
+    def put(self, request, id):
         token = request.headers.get('Authorization', None)
         payload = LoginSerializer.validateToken(token)
-        author = Author.objects.get(pk = payload.get('user_id', None))
-        post = Post.objects.get(pk = pk)
+        author = Author.objects.get(id = payload.get('user_id', None))
+        post = Post.objects.get(id = id)
         if post.author == author:
             serializer = PostSerializer(post, data=request.data)
             if serializer.is_valid():
@@ -84,11 +84,11 @@ class PostViewSet(APIView):
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
         return Response(status=status.HTTP_401_UNAUTHORIZED)  
     
-    def delete(self, request, pk):
+    def delete(self, request, id):
         token = request.headers.get('Authorization', None)
         payload = LoginSerializer.validateToken(token)
-        author = Author.objects.get(pk = payload.get('user_id', None))
-        post = Post.objects.get(pk = pk)
+        author = Author.objects.get(id = payload.get('user_id', None))
+        post = Post.objects.get(id = id)
         if post.author == author:
             post.delete()
             return Response(status=status.HTTP_204_NO_CONTENT)
@@ -110,11 +110,11 @@ class LoginView(APIView):
 class FriendRequestViewSet(APIView):
  
     def get(self, request, *args, **kwargs):
-        if (kwargs.get('pk') and not kwargs.get("fk") ): #/api/friendrequest/OBJECT(pk = OBJECT/RECIEVER)
-            serializer = FriendRequestSerializer(FriendRequest.objects.filter(object__id=kwargs.get('pk')),many=True)
+        if (kwargs.get('id') and not kwargs.get("fk") ): #/api/friendrequest/OBJECT(id = OBJECT/RECIEVER)
+            serializer = FriendRequestSerializer(FriendRequest.objects.filter(object__id=kwargs.get('id')),many=True)
             return Response(serializer.data, status=status.HTTP_200_OK)
-        elif (kwargs.get("pk") and kwargs.get("fk")): #/api/ACTOR/friendrequest/OBJECT (pk = OBJECT/RECIEVER)
-            filter_data = FriendRequest.objects.filter(actor__id=kwargs.get('pk'))
+        elif (kwargs.get("id") and kwargs.get("fk")): #/api/ACTOR/friendrequest/OBJECT (id = OBJECT/RECIEVER)
+            filter_data = FriendRequest.objects.filter(actor__id=kwargs.get('id'))
             filter_data = filter_data.filter(object__id = kwargs.get('fk'))
             serializer = FriendRequestSerializer(filter_data,many=True)
 
@@ -134,26 +134,26 @@ class FriendRequestViewSet(APIView):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
     
     def delete(self,request,*args,**kwargs):
-        if (kwargs.get('pk')and kwargs.get("fk") ): #/api/ACTOR/friendrequest/OBJECT 
-            filter_data = FriendRequest.objects.filter(actor__id=kwargs.get('pk'))
+        if (kwargs.get('id')and kwargs.get("fk") ): #/api/ACTOR/friendrequest/OBJECT 
+            filter_data = FriendRequest.objects.filter(actor__id=kwargs.get('id'))
             filter_data = filter_data.filter(object__id = kwargs.get('fk'))
             filter_data.delete()
-            serializer = FriendRequestSerializer(FriendRequest.objects.filter(actor__id=kwargs.get('pk')),many=True)
+            serializer = FriendRequestSerializer(FriendRequest.objects.filter(actor__id=kwargs.get('id')),many=True)
          
         return Response(status=status.HTTP_204_NO_CONTENT)
 
 class FollowersViewSet(APIView):
    def get(self, request, *args, **kwargs):
-        if kwargs.get("pk") and not kwargs.get("fk"):
-           pk_val = kwargs.get("pk")
-           serializer =  FollowersSerializer(Followers.objects.filter(user__id=pk_val),many=True) 
+        if kwargs.get("id") and not kwargs.get("fk"):
+           id_val = kwargs.get("id")
+           serializer =  FollowersSerializer(Followers.objects.filter(user__id=id_val),many=True) 
         elif kwargs.get("fk"):
-            pk_val = kwargs.get("pk")
+            id_val = kwargs.get("id")
             fk_val = kwargs.get("fk")
-            serializer =  FollowersSerializer(Followers.objects.filter(user__id=pk_val),many=True) 
+            serializer =  FollowersSerializer(Followers.objects.filter(user__id=id_val),many=True) 
            
             if fk_val in serializer.data[0]["items"]:
-                serializer =  FollowersSerializer(Followers.objects.filter(user__id=pk_val),many=True) 
+                serializer =  FollowersSerializer(Followers.objects.filter(user__id=id_val),many=True) 
                 return Response(serializer.data, status=status.HTTP_200_OK)
             else:
                 return Response("Not a follower",status=status.HTTP_200_OK)
@@ -167,35 +167,73 @@ class FollowersViewSet(APIView):
    def put(self, request, **kwargs):
         token = request.headers.get('Authorization', None)
 
-        if kwargs.get("pk") and not kwargs.get("fk"):
-           pk_val = kwargs.get("pk")
+        if kwargs.get("id") and not kwargs.get("fk"):
+           id_val = kwargs.get("id")
            new_follower = request.data.get("items")
     
-           followers =Followers.objects.get(user__id=pk_val)
+           followers =Followers.objects.get(user__id=id_val)
            followers.items.add(new_follower)
 
-           serializer =  FollowersSerializer(Followers.objects.filter(user__id=pk_val),many=True) 
+           serializer =  FollowersSerializer(Followers.objects.filter(user__id=id_val),many=True) 
            
 
         return Response(serializer.data, status=status.HTTP_200_OK)
    
    def delete(self,request,*args, **kwargs ):
-    pk_val = kwargs.get("pk")
+    id_val = kwargs.get("id")
     fk_val = kwargs.get("fk")
-    serializer =  FollowersSerializer(Followers.objects.filter(user__id=pk_val),many=True) 
-    new_list = Followers.objects.get(user__id=pk_val)
+    serializer =  FollowersSerializer(Followers.objects.filter(user__id=id_val),many=True) 
+    new_list = Followers.objects.get(user__id=id_val)
     new_list.items.remove(fk_val)
     
     
 
     return Response(serializer.data, status=status.HTTP_200_OK)
     
+
+
+# proper format
+@api_view(['GET', 'POST', 'PUT', 'DELETE'])
 def authors(request, author_id = None):
-    authors = Author.objects.all()
-    if author_id:
-        authors = authors.filter(pk = author_id)
-    data = serializers.serialize('json', authors)
-    return JsonResponse(data, safe=False)
+    if request.method == 'GET':
+        authors = Author.objects.all()
+        if author_id:
+            authors = authors.filter(id = author_id)
+        data = serializers.serialize('json', authors)
+        return JsonResponse(data, safe=False)
+    elif request.method == 'POST':
+        data = json.parse(request)
+        serializer = AuthorSerializer(data=data)
+        if serializer.is_valid():
+            serializer.save()
+            # create inboxItems for new author
+            inboxItem = InboxItem.objects.create(author = serializer.data.get('id'))
+            inboxItem.save()
+            # create inbox for new author
+            inbox = Inbox.objects.create(author = serializer.data.get('id'))
+            inbox.save()
+            return JsonResponse(serializer.data, status=status.HTTP_201_CREATED)
+        return JsonResponse(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    elif request.method == 'PUT':
+        data = json.parse(request)
+        token = request.headers.get('Authorization', None)
+        payload = LoginSerializer.validateToken(token)
+        author = Author.objects.get(id = payload.get('user_id', None))
+        if author_id != author.id:
+            return JsonResponse("Not authorized", status=status.HTTP_401_UNAUTHORIZED, safe=False)
+        serializer = AuthorSerializer(author, data=data)
+        if serializer.is_valid():
+            serializer.save()
+            return JsonResponse(serializer.data, status=status.HTTP_200_OK)
+        return JsonResponse(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    elif request.method == 'DELETE':
+        token = request.headers.get('Authorization', None)
+        payload = LoginSerializer.validateToken(token)
+        author = Author.objects.get(id = payload.get('user_id', None))
+        if author_id != author.id:
+            return JsonResponse("Not authorized", status=status.HTTP_401_UNAUTHORIZED, safe=False)
+        author.delete()
+        return JsonResponse("Deleted", status=status.HTTP_204_NO_CONTENT, safe=False)
 
 def followers(request, author_id = None):
     # not implemented
@@ -209,27 +247,56 @@ def friends(request, author_id):
     # not implemented
     return JsonResponse("Not implemented", status=status.HTTP_501_NOT_IMPLEMENTED, safe=False)
 
+@api_view(['GET', 'POST', 'PUT', 'DELETE'])
 def posts(request, author_id, post_id = None):
     if request.method == 'GET':
-        author = Author.objects.get(pk = author_id)
+        author = Author.objects.get(id = author_id)
         posts = Post.objects.filter(author= author)
+        token = request.headers.get('Authorization', None)
+        try:
+            payload = LoginSerializer.validateToken(token)
+        except:
+            payload = None
+        if payload:
+            requestAuthor = Author.objects.get(id = payload.get('user_id', None))
+            if author_id != str(requestAuthor.id):
+                posts = posts.filter(visibility = 'PUBLIC')
+
         if post_id:
-            posts = posts.filter(pk = post_id)
-        data = serializers.serialize('json', posts)
-        return JsonResponse(data, safe=False)
+            posts = posts.filter(id = post_id)
+            posts = PostSerializer(posts, many=True).data
+            posts[0]['author'] = AuthorSerializer(Author.objects.get(id = posts[0]['author'])).data
+            return Response(posts[0], status=status.HTTP_200_OK)
+        
+        serializer = PostSerializer(posts, many=True)
+        for post in serializer.data:
+            post['author'] = AuthorSerializer(Author.objects.get(id = post['author'])).data
+
+        return Response(serializer.data, status=status.HTTP_200_OK)
+    
+
     elif request.method == 'POST':
-        author = Author.objects.get(pk = author_id)
+        token = request.headers.get('Authorization', None)
+        payload = LoginSerializer.validateToken(token)
+        author = Author.objects.get(id = payload.get('user_id', None))
+        if author_id != str(author.id):
+            return JsonResponse("Not authorized", status=status.HTTP_401_UNAUTHORIZED, safe=False)
         data = request.data
         data['author'] = author.id
         data['authorName'] = author.displayName
-        data = PostSerializer(data=data)
+        serializer = PostSerializer(data=data)
         if serializer.is_valid():
             serializer.save()
             return JsonResponse(serializer.data, status=status.HTTP_201_CREATED, safe=False)
         return JsonResponse(serializer.errors, status=status.HTTP_400_BAD_REQUEST, safe=False)
+    
     elif request.method == 'PUT':
-        author = Author.objects.get(pk = author_id)
-        post = Post.objects.get(pk = post_id)
+        token = request.headers.get('Authorization', None)
+        payload = LoginSerializer.validateToken(token)
+        author = Author.objects.get(id = payload.get('user_id', None))
+        if author_id != str(author.id):
+            return JsonResponse("Not authorized", status=status.HTTP_401_UNAUTHORIZED, safe=False)
+        post = Post.objects.get(id = post_id)
         if post.author == author:
             serializer = PostSerializer(post, data=request.data)
             if serializer.is_valid():
@@ -237,9 +304,14 @@ def posts(request, author_id, post_id = None):
                 return JsonResponse(serializer.data, status=status.HTTP_200_OK, safe=False)
             return JsonResponse(serializer.errors, status=status.HTTP_400_BAD_REQUEST, safe=False)
         return JsonResponse("Not authorized", status=status.HTTP_401_UNAUTHORIZED, safe=False)
+    
     elif request.method == 'DELETE':
-        author = Author.objects.get(pk = author_id)
-        post = Post.objects.get(pk = post_id)
+        token = request.headers.get('Authorization', None)
+        payload = LoginSerializer.validateToken(token)
+        author = Author.objects.get(id = payload.get('user_id', None))
+        if author_id != str(author.id):
+            return JsonResponse("Not authorized", status=status.HTTP_401_UNAUTHORIZED, safe=False)
+        post = Post.objects.get(id = post_id)
         if post.author == author:
             post.delete()
             return JsonResponse("Deleted", status=status.HTTP_204_NO_CONTENT, safe=False)
@@ -252,7 +324,7 @@ def comments(request, author_id, post_id):
         data = serializers.serialize('json', comments)
         return JsonResponse(data, safe=False)
     elif request.method == 'POST':
-        author = Author.objects.get(pk = author_id)
+        author = Author.objects.get(id = author_id)
         data = request.data
         data['author'] = author.id
         data['authorName'] = author.displayName
@@ -277,19 +349,24 @@ def authorFollowersPosts(request, author_id):
 def inbox(request, author_id):
     if request.method == 'GET':
         # get all inbox items for author
-        author = Author.objects.get(pk = author_id)
+        author = Author.objects.get(id = author_id)
         inbox = Inbox.objects.filter(author = author)
         serializer = InboxSerializer(inbox, many=True)
         # get InboxItem from data['items']
-        serializer.data[0]['items'] = InboxItemSerializer(InboxItem.objects.get(pk = serializer.data[0]['items'])).data
+        serializer.data[0]['items'] = InboxItemSerializer(InboxItem.objects.get(id = serializer.data[0]['items'])).data
         # get Post from data['items']['post']
         for i in range(len(serializer.data[0]['items']['posts'])):
-            serializer.data[0]['items']['posts'][i] = PostSerializer(Post.objects.get(pk = serializer.data[0]['items']['posts'][i])).data
+            serializer.data[0]['items']['posts'][i] = PostSerializer(Post.objects.get(id = serializer.data[0]['items']['posts'][i])).data
         for i in range(len(serializer.data[0]['items']['requests'])):
-            serializer.data[0]['items']['requests'][i] = PostSerializer(Post.objects.get(pk = serializer.data[0]['items']['requests'][i])).data
+            serializer.data[0]['items']['requests'][i] = RequestSerializer(Request.objects.get(id = serializer.data[0]['items']['requests'][i])).data
+        for i in range(len(serializer.data[0]['items']['comments'])):
+            serializer.data[0]['items']['comments'][i] = CommentSerializer(Comment.objects.get(id = serializer.data[0]['items']['comments'][i])).data
+        for i in range(len(serializer.data[0]['items']['likes'])):
+            serializer.data[0]['items']['likes'][i] = LikeSerializer(Like.objects.get(id = serializer.data[0]['items']['likes'][i])).data
+
 
         # items becomes a list with all posts and requests
-        serializer.data[0]['items'] = serializer.data[0]['items']['posts'] + serializer.data[0]['items']['requests']
+        serializer.data[0]['items'] = serializer.data[0]['items']['posts'] + serializer.data[0]['items']['requests'] + serializer.data[0]['items']['comments'] + serializer.data[0]['items']['likes']
         
         return JsonResponse(serializer.data[0], safe=False)
         
@@ -299,7 +376,7 @@ def inbox(request, author_id):
     
     elif request.method == 'DELETE':
         # clear the items field of the inbox with author_id = author_id
-        author = Author.objects.get(pk = author_id)
+        author = Author.objects.get(id = author_id)
         inbox = Inbox.objects.get(author = author)
         inbox.items.posts.clear()
         inbox.items.requests.clear()
